+++ backend/volume/src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
-import { UserControllerModule } from './controllers/user/userController.module';
 import { PrismaModule } from './prisma/prisma.module';
-
+import { UserController } from './user/user.controller';
+import { UserModule } from './user/user.module'
 
 @Module({
-  imports: [PrismaModule, UserControllerModule],
+  imports: [PrismaModule, UserModule],
   controllers: [AppController],
   providers: [],
 })
