+++ backend/volume/src/game/game.controller.ts
@@ -5,7 +5,7 @@ import {
   Post,
 } from '@nestjs/common';
 import { Game as GameModel} from '@prisma/client';
-import { PrismaGameService } from 'src/prisma/game/prismaGame.service';
+import { PrismaGameService } from './prisma/prismaGame.service';
 
 // should only be allowed to get games as these will be created by the engine
 @Controller('game')
