+++ backend/volume/src/game/game.module.ts
@@ -1,14 +1,13 @@
 import { Module } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { PrismaGameModule } from 'src/prisma/game/prismaGame.module';
-import { PrismaGameService } from 'src/prisma/game/prismaGame.service';
-import { GameController } from './gameController.controller';
+import { PrismaGameService } from './prisma/prismaGame.service';
+import { GameController } from './game.controller';
 
 @Module({
-  imports: [PrismaClient, PrismaGameModule],
+  imports: [PrismaClient],
   controllers: [GameController],
   providers: [PrismaGameService, PrismaService, GameController],
   exports: [GameController],
 })
-export class GameControllerModule {}
+export class GameModule {}
