+++ backend/volume/src/msg/prismaMsg.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
-import { PrismaService } from '../prisma.service';
+import { PrismaService } from 'src/prisma/prisma.service';
 import { PrismaMsgService } from './prismaMsg.service';
 
 @Module({
