+++ backend/volume/src/msg/prismaMsg.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from '../prisma.service';
+import { PrismaService } from 'src/prisma/prisma.service';
 import { Msg, Prisma } from '@prisma/client';
 
 @Injectable()
